Sua template & them chuc nang danh gia sao cho san pham

+++ app/database/migrations/2019_05_01_025945_create_reviews_table.php
@@ -16,7 +16,7 @@ class CreateReviewsTable extends Migration
         Schema::create('reviews', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->timestamps();
-            $table->tinyInteger('stars');
+            $table->integer('stars');
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('user_id');
             $table->text('content');
